// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const EmployeeStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE'
} as const

export type EmployeeStatus = (typeof EmployeeStatus)[keyof typeof EmployeeStatus]


export const Role = {
  ADMIN: 'ADMIN',
  COMMUNICATION_MANAGER: 'COMMUNICATION_MANAGER',
  HR_MANAGER: 'HR_MANAGER',
  MANAGER: 'MANAGER',
  EMPLOYEE: 'EMPLOYEE'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const PublicationType = {
  ANNOUNCEMENT: 'ANNOUNCEMENT',
  NEWS: 'NEWS'
} as const

export type PublicationType = (typeof PublicationType)[keyof typeof PublicationType]


export const PublicationStatus = {
  DRAFT: 'DRAFT',
  SCHEDULED: 'SCHEDULED',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED',
  EXPIRED: 'EXPIRED'
} as const

export type PublicationStatus = (typeof PublicationStatus)[keyof typeof PublicationStatus]


export const AudienceType = {
  ALL: 'ALL',
  UNIT: 'UNIT',
  DEPARTMENT: 'DEPARTMENT',
  GROUP: 'GROUP',
  USER: 'USER'
} as const

export type AudienceType = (typeof AudienceType)[keyof typeof AudienceType]


export const EventStatus = {
  DRAFT: 'DRAFT',
  SCHEDULED: 'SCHEDULED',
  PUBLISHED: 'PUBLISHED',
  CANCELLED: 'CANCELLED',
  FINISHED: 'FINISHED'
} as const

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus]


export const DocumentStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  SUPERSEDED: 'SUPERSEDED',
  ARCHIVED: 'ARCHIVED'
} as const

export type DocumentStatus = (typeof DocumentStatus)[keyof typeof DocumentStatus]